:refactor: change component name
